feat: set up email package data class add type hinting for method send in class EmailSender move function load_recipients() to file_worker module
