+++ src/bmailer/__init__.py
@@ -1,0 +1,5 @@
+from .modules.sender import EmailSender
+
+__all__ = [
+    "EmailSender",
+]
