+++ src/bmailer/models/email_package.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class EmailPackage:
+    to: str
+    data: str
